pyproject.toml
[project]
name = "podcast-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "cozepy>=0.19.0",
    "langchain-community>=0.3.27",
    "langchain-unstructured>=0.1.6",
    "modelscope[audio]>=1.24.1",
    "numpy>=2.2.6",
    "python-dotenv>=1.1.1",
    "unstructured>=0.18.5",
]
